requirements.txt
pytrends
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
fastapi
uvicorn
python-multipart
apscheduler
requests
beautifulsoup4
feedparser
markdown
python-telegram-bot>=21.0
lxml
anthropic
Pillow
requests-oauthlib
claude-agent-sdk
# Phase 2 (쇼츠 변환)
gTTS
openai
pydub
# Phase 2 (YouTube 업로드 진행 표시)
google-resumable-media
# Phase 3 (엔진 추상화 — 선택적 의존성)
# google-generativeai  # Gemini Writer / Veo 사용 시 pip install google-generativeai
